(bug) manually add the orchestrator class to the replica class list
The puppet_enterprise::profile::orchestrator class is not declared
on the Replica until the Replica is promoted.

With this commit ...

Manually add the puppet_enterprise::profile::orchestrator class
to the list of classes on the Replica to ensure that the
Master and Replica receive the same set of settings.

(This does not actually apply or declare the class on the Replica.)